Move rotate_image_by_angle() to utils
