+++ scripts/extract.py
@@ -9,12 +9,13 @@ from tqdm import tqdm
 
 from lib.gpu_stats import GPUStats
 from lib.multithreading import pool_process
+from lib.utils import rotate_image_by_angle
 from scripts.fsmedia import Alignments, Faces, Images, Utils
 
 tqdm.monitor_interval = 0  # workaround for TqdmSynchronisationWarning
 
 
-class Extract(object):
+class Extract():
     """ The extract process. """
 
     def __init__(self, arguments):
@@ -27,7 +28,9 @@ class Extract(object):
         self.output_dir = self.faces.output_dir
 
         self.export_face = True
-        self.save_interval = self.args.save_interval if hasattr(self.args, "save_interval") else None
+        self.save_interval = None
+        if hasattr(self.args, "save_interval"):
+            self.save_interval = self.args.save_interval
 
     def process(self):
         """ Perform the extraction process """
@@ -53,8 +56,8 @@ class Extract(object):
         self.faces.num_faces_detected = faces
 
     def write_alignments(self):
+        """ Save the alignments file """
         self.alignments.write_alignments(self.faces.faces_detected)
-
 
     def extract_single_process(self):
         """ Run extraction in a single process """
@@ -94,7 +97,7 @@ class Extract(object):
             image = Utils.cv2_read_write('read', filename)
 
             for angle in self.images.rotation_angles:
-                currentimage = Utils.rotate_image_by_angle(image, angle)
+                currentimage = rotate_image_by_angle(image, angle)
                 faces = self.faces.get_faces(currentimage, angle)
                 process_faces = [(idx, face) for idx, face in faces]
                 if process_faces and angle != 0 and self.args.verbose:
